+++ docs/fr_FR/Changelog.rst
@@ -2,6 +2,8 @@ ChangeLog
 =========
 
 - Page de l'équipement: Corrections translation 'en_US'.
+- Interne: jeedom.eqLogic.builSelectCmd() => buildSelectCmd(). Core v4.0 min.
+- Interne: Mise-à-jour AbeilleNoise.
 
 230730-BETA-1
 -------------